Cast LARAVEL_START const to float (#1968)
* Cast LARAVEL_START const to float

* Cast LARAVEL_START const to float (2)

---------

+++ src/LaravelDebugbar.php
@@ -97,7 +97,7 @@ class LaravelDebugbar extends DebugBar
 
     public function __construct(Application $app, Request $request)
     {
-        $startTime = defined('LARAVEL_START') ? LARAVEL_START : microtime(true);
+        $startTime = defined('LARAVEL_START') ? (float) LARAVEL_START : microtime(true);
 
         $this->app = $app;
         $this->request = $request;
@@ -215,7 +215,7 @@ class LaravelDebugbar extends DebugBar
 
     public function booted(): void
     {
-        $startTime = defined('LARAVEL_START') ? LARAVEL_START : null;
+        $startTime = defined('LARAVEL_START') ? (float) LARAVEL_START : null;
         if ($startTime) {
             $this->addMeasure('Booting', $startTime, microtime(true));
         }
